bugfix on unresetting component state on explicit close on the uploader
