--- components/ReferenceManager/references/reference_uploader/referenceDefaultSchemas.ts
@@ -1,1 +1,0 @@
-export const DEFAULT_SCHEMA_TYPE = "";
