+++ components/ReferenceManager/references/reference_uploader/reference_default_schemas.ts
@@ -1,0 +1,11 @@
+export type ReferenceSchemaValueSet = {
+  attachment: File | null;
+  schema: any;
+  required: string[];
+};
+
+export const DEFAULT_REF_SCHEMA_SET: ReferenceSchemaValueSet = {
+  attachment: null,
+  schema: {},
+  required: [],
+};
